Modifier keys work

+++ VirtualMouseInteractor/InputMessage.cs
@@ -15,9 +15,9 @@ namespace VirtualDeviceInteractor
         public readonly int buttons;
         public readonly int key;
         public readonly int keyState;
-        public readonly int modifiers;
+        public readonly byte modifiers;
 
-        public InputMessage(char xAxis = '\0', char yAxis = '\0', int buttons = 0, int key = 0, int keyState = 0, int modifiers = 0)
+        public InputMessage(char xAxis = '\0', char yAxis = '\0', int buttons = 0, int key = 0, int keyState = 0, byte modifiers = 0)
         {
             this.xAxis = xAxis;
             this.yAxis = yAxis;
